refactor(subscriber-enrollment): replace `GeneratedEmployee` with `SubscriberEnrollment`, restructure package to `fakeData`, and update references
