+++ src/main/java/com/fastChickensHR/edi/fakeData/x834ContextGenerator.java
@@ -5,7 +5,7 @@
  *
  * For license information see the LICENSE file in the root of this project.
  */
-package com.fastChickensHR.edi.testdata;
+package com.fastChickensHR.edi.fakeData;
 
 import com.fastChickensHR.edi.x834.x834Context;
 import net.datafaker.Faker;
