+++ src/main/java/com/fastChickensHR/edi/x834/enrollment/SubscriberEnrollment.java
@@ -5,7 +5,7 @@
  *
  * For license information see the LICENSE file in the root of this project.
  */
-package com.fastChickensHR.edi.testdata;
+package com.fastChickensHR.edi.x834.enrollment;
 
 import com.fastChickensHR.edi.x834.loop2000.Member;
 import com.fastChickensHR.edi.x834.loop2000.loop3000.HealthCoverage;
@@ -13,8 +13,11 @@ import com.fastChickensHR.edi.x834.loop2000.loop3000.HealthCoverage;
 import java.util.List;
 
 /**
- * Holder for a generated subscriber {@link Member} together with the
+ * Aggregate representing the enrollment data required to build an 834 file
+ * for a single subscriber: the subscriber {@link Member} together with the
  * {@link HealthCoverage} (HD) segments that apply to that employee/family.
+ *
+ * <p>This is the canonical input contract passed into the x834 builder.
  */
-public record GeneratedEmployee(Member member, List<HealthCoverage> coverages) {
+public record SubscriberEnrollment(Member member, List<HealthCoverage> coverages) {
 }
